divide components in main.js to modules #1
