+++ src/components/cost-template.js
@@ -1,0 +1,9 @@
+const createCostTemplate = () => {
+  return (
+    `<p class="trip-info__cost">
+      Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
+    </p>`
+  );
+};
+
+export {createCostTemplate};
