+++ src/components/trip-days-list-template.js
@@ -1,0 +1,7 @@
+const createTripDaysListTemplate = () => {
+  return (
+    `<ul class="trip-days"></ul>`
+  );
+};
+
+export {createTripDaysListTemplate};
