+++ src/components/trip-events-list-template.js
@@ -1,0 +1,7 @@
+const createTripEventsListTemplate = () => {
+  return (
+    `<ul class="trip-events__list"></ul>`
+  );
+};
+
+export {createTripEventsListTemplate};
